resolve bcrypt
