+++ package.json
@@ -12,7 +12,7 @@
   },
   "name": "jonathan-silva-capstone-server",
   "version": "1.0.0",
-  "main": "index.js",
+  "main": "server.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "node --watch server.js"
